more reorganizing + msg in Favorite.js + PropertyItemgit status
