+++ client/src/components/FavoriteItem/index.js
@@ -1,60 +1,64 @@
-import React from "react";
-import { Link } from "react-router-dom";
-import { useStoreContext } from "../../utils/GlobalState";
-import { ADD_TO_FAVORITES, UPDATE_FAVORITE_QUANTITY } from "../../utils/actions";
+import React, { useState } from 'react';
+import { useDispatch } from 'react-redux';
+import { REMOVE_FROM_FAVORITES, UPDATE_FAVORITE_QUANTITY } from "../../utils/actions";
 import { idbPromise } from "../../utils/helpers";
 
-function FavoriteItem(item) {
-  const [state, dispatch] = useStoreContext();
+const FavoriteItem = ({ item }) => {
+  const dispatch = useDispatch();
+  const [isFavorited, setIsFavorited] = useState(true);
 
-  const {
-    image,
-    name,
-    _id,
-    price,
-  } = item;
-
-  const { property } = state
-
-const addToFavorites = () => {
-  const itemInFavorites = property.find((propertyItem) => propertyItem._id === _id)
-  if (itemInFavorites) {
+  const removeFromFavorites = () => {
     dispatch({
-      type: UPDATE_FAVORITE_QUANTITY,
-      _id: _id,
-      favoriteQuantity: parseInt(itemInFavorites.favoriteQuantity) + 1
+      type: REMOVE_FROM_FAVORITES,
+      _id: item._id
     });
-    idbPromise('favorites', 'put', {
-      ...itemInFavorites,
-      favoriteQuantity: parseInt(itemInFavorites.favoriteQuantity) + 1
-    });
-  } else {
-    dispatch({
-      type: ADD_TO_FAVORITES,
-      product: { ...item, purchaseQuantity: 1 }
-    });
-    idbPromise('favorites', 'put', { ...item, favoriteQuantity: 1 });
+    idbPromise('favorite', 'delete', { ...item });
+    setIsFavorited(false);
+  };
+
+  const onChange = (e) => {
+    const value = e.target.value;
+    if (value === '0') {
+      removeFromFavorites();
+    } else if (!isFavorited) {
+      dispatch({
+        type: UPDATE_FAVORITE_QUANTITY,
+        _id: item._id,
+        favoriteQuantity: parseInt(value)
+      });
+      idbPromise('favorite', 'put', { ...item, favoriteQuantity: parseInt(value) });
+      setIsFavorited(true);
+    }
   }
-}
 
-return (
-  <div className="card px-1 py-1">
-    <Link to={`/property/${_id}`}>
-      <img
-        alt={name}
-        src={`/images/${image}`}
-      />
-      <p>{name}</p>
-    </Link>
-    <div>
-      <span>${price}</span>
-
-      {/* ADD TO PROPERTY.JS */}
+  return (
+    <div className="flex-row">
+      <div>
+        <img
+          src={`/images/${item.image}`}
+          alt=""
+        />
+      </div>
+      <div>
+        <div>{item.name}, ${item.price}</div>
+        <div>
+          <input
+            type="number"
+            placeholder="0"
+            value={item.favoriteQuantity}
+            onChange={onChange}
+          />
+          <span
+            role="img"
+            aria-label="trash"
+            onClick={() => removeFromFavorites(item)}
+          >
+            ❌
+          </span>
+        </div>
+      </div>
     </div>
-    <button onClick={addToFavorites}>Add to Favorites</button>
-  </div>
- 
-);
+  );
 }
 
 export default FavoriteItem;
