+++ package.json
@@ -19,6 +19,7 @@
   "dependencies": {
     "apollo-server-express": "^3.11.1",
     "express": "^4.17.2",
-    "graphql": "^15.8.0"
+    "graphql": "^15.8.0",
+    "react-redux": "^8.1.1"
   }
 }
